refactor: Rename table and columns for clarity
Renamed the final table in the database from `t_vacances` to `t_vacances_scolaires` to better reflect its content.

Also, the columns `date` and `annee_scolaire` in the final table and the corresponding DataFrame are now prefixed with `vac_` for better clarity and to avoid potential naming conflicts. This includes:

- `date` is now `vac_date`
- `annee_scolaire` is now `vac_annee_scolaire`

The ETL pipeline and the associated test notebook have been updated to reflect these changes. The code for dropping the staging table has been removed from the test notebook.